requirements.txt
fastapi~=0.111.0
SQLAlchemy~=2.0.30
python-dotenv~=1.0.1
pydantic~=2.7.1
bcrypt~=4.1.3
PyJWT~=2.8.0
starlette~=0.37.2
python-socketio~=5.11.0
uvicorn~=0.29.0